Give a subtle visual hint (gray box-shadow) as to which areas on the screen can be edited
Divided .episode_list into .header and .the_list sections so that we
could put some space under .header.

Made it so that the context_ribbon for the #new_k3cms_s3_podcast_episode
is initially visible when you click New episode and so that it doesn't
go away when the mouse leaves that trigger element, as it would normally
for pre-existing records.

Fixed bug that occured when you submited two new episodes in a row: The
old .context_ribbon.new_k3cms_s3_podcast_episode was hanging around and
interfering.